+++ stanza/example_pingstream_test.go
@@ -18,7 +18,7 @@ import (
 // a ping payload.
 func WrapPingIQ(to jid.JID) xml.TokenReader {
 	start := xml.StartElement{Name: xml.Name{Local: "ping", Space: "urn:xmpp:ping"}}
-	return stanza.WrapIQ(stanza.IQ{To: to, Type: stanza.GetIQ}, xmlstream.Wrap(nil, start))
+	return stanza.IQ{To: to, Type: stanza.GetIQ}.Wrap(xmlstream.Wrap(nil, start))
 }
 
 func Example_stream() {